requirements.txt
fastapi[standard]
gspread
gspread-formatting
requests
python-dotenv
gunicorn
telethon
SQLAlchemy
